feat: Update BettingPanel and ActivePlayersList components
