+++ public/index.html
@@ -7,7 +7,7 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="Aviator - The ultimate crash game experience. Place bets, watch the multiplier soar, and cash out before the plane crashes! Real-time multiplayer gaming with provably fair results."
     />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
@@ -24,7 +24,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>🚀 Aviator - Crash Game</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
